requirements.txt
atlassian-python-api==3.22.0
beautifulsoup4==4.10.0
pytest==7.1.2
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0